# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SkyVoyage Weather + Travel Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /weather/current:
    get:
      operationId: getCurrentWeather
      tags: [weather]
      summary: Get current weather for a location
      parameters:
        - in: query
          name: city
          schema: { type: string }
          description: City name to search
      responses:
        "200":
          description: Current weather data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentWeather"
  /weather/hourly:
    get:
      operationId: getHourlyForecast
      tags: [weather]
      summary: Get hourly weather forecast
      parameters:
        - in: query
          name: city
          schema: { type: string }
      responses:
        "200":
          description: Hourly forecast
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HourlyForecast"
  /weather/weekly:
    get:
      operationId: getWeeklyForecast
      tags: [weather]
      summary: Get 7-day weather forecast
      parameters:
        - in: query
          name: city
          schema: { type: string }
      responses:
        "200":
          description: 7-day forecast
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyForecast"

  # ─── DESTINATIONS ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    CurrentWeather:
      type: object
      required: [city, country, temperature, condition, humidity, windSpeed, uvIndex, sunrise, sunset, feelsLike, visibility, airQuality]
      properties:
        city: { type: string }
        country: { type: string }
        temperature: { type: number }
        condition: { type: string }
        conditionCode: { type: string }
        humidity: { type: integer }
        windSpeed: { type: number }
        uvIndex: { type: number }
        sunrise: { type: string }
        sunset: { type: string }
        feelsLike: { type: number }
        visibility: { type: number }
        airQuality: { type: integer }
        airQualityLabel: { type: string }
        precipitation: { type: number }
    HourlyForecast:
      type: object
      required: [hour, temperature, condition, precipitation]
      properties:
        hour: { type: string }
        temperature: { type: number }
        condition: { type: string }
        conditionCode: { type: string }
        precipitation: { type: number }
        windSpeed: { type: number }
    DailyForecast:
      type: object
      required: [date, dayName, high, low, condition, precipitation]
      properties:
        date: { type: string }
        dayName: { type: string }
        high: { type: number }
        low: { type: number }
        condition: { type: string }
        conditionCode: { type: string }
        precipitation: { type: number }
        humidity: { type: integer }

    # ─── DESTINATIONS ─────────────────────────────────────────────────────────
...