# lib/api-zod/src/generated/types/currentWeather.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SkyVoyage Weather + Travel Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface CurrentWeather {
  city: string;
  country: string;
  temperature: number;
  condition: string;
  conditionCode?: string;
  humidity: number;
  windSpeed: number;
  uvIndex: number;
  sunrise: string;
  sunset: string;
  feelsLike: number;
  visibility: number;
  airQuality: number;
  airQualityLabel?: string;
  precipitation?: number;
}